Move content block directives to separate file
